+++ src/test/java/com/societegenerale/aut/main/ObjectWithJavaTextDateFormat.java
@@ -1,9 +1,0 @@
-package com.societegenerale.aut.main;
-
-import java.text.DateFormat;
-
-public class ObjectWithJavaTextDateFormat {
-
-	private DateFormat dateFormat;
-
-}