+++ src/test/java/com/societegenerale/aut/test/TestClassWhichThrowGenericException.java
@@ -1,8 +1,0 @@
-package com.societegenerale.aut.test;
-
-public class TestClassWhichThrowGenericException {
-
-    public void testMethod() throws Exception{
-        throw new Exception("For testing custom rules");
-    }
-}