+++ src/test/java/com/societegenerale/aut/test/specificCase/DummyClassToValidate.java
@@ -1,4 +1,0 @@
-package com.societegenerale.aut.test.specificCase;
-
-public class DummyClassToValidate {
-}