+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 ### Changed
+- PR #30 - Now using https://github.com/societe-generale/arch-unit-build-plugin-core as the foundation for the maven plugin 
 - PR #20 - using Maven recommended way for logging - thanks [@croesh](https://github.com/croesh) for the contribution !
 - PR #27 - setting Maven plugins versions to avoid warning at build time - thanks [@khmarbaise](https://github.com/khmarbaise) for the contribution !